+++ source/Core/Edition/EditionPathProvider.php
@@ -27,6 +27,8 @@ namespace OxidEsales\EshopCommunity\Core\Edition;
  *
  * @internal Do not make a module extension for this class.
  * @see      http://oxidforge.org/en/core-oxid-eshop-classes-must-not-be-extended.html
+ *
+ * @deprecated since v6.0.0-rc.2 (2017-08-24); Use \OxidEsales\Facts\Facts instead.
  */
 class EditionPathProvider
 {